[CUERipper] Save metadata before UpdateDrive(s)
Up to now, entered metadata was cleared and lost, when UpdateDrive() or
UpdateDrives() was called after device change events. This happened for
example in case of a DBT_DEVNODES_CHANGE event. Typical triggers for
this event are often not even CD drive related, like insert/eject of
USB sticks, disconnection or connection of a Bluetooth mouse etc.

- In order to avoid loss of entered metadata in case of such events,
  save the data first before calling the methods UpdateDrive()
  or UpdateDrives().
- The entered metadata is now also saved, when a CD is ejected.
- Fixes #20