[CUETools] Fix CTDB.FetchFile Exception
If "Verify using CTDB" is disabled, an exception can occur in
LoadAndResizeAlbumArt(). CTDB is null in this case and
CTDB.FetchFile(imageMeta.uri, ms) will throw the exception.
The exception occurs after selecting specific entries in the
"Select best match" window for albums, when there are different album
arts available from coverartarchive.org. In this case, the window for
selecting the album art is not shown anymore and the exception appears.

- Use "try" and catch the exception
- Fixes the following error:
  Exception Object reference not set to an instance of an object.
- This commit is an addition to 3f121a8
- Resolves #148