Add freedb site address setting
Until now, the address for the freedb site has been hard-coded into
CUETools and in case of CUERipper it has been an application
setting (MAIN_FREEDB_SITEADDRESS stored in CUERipper.exe.config).
The current default address for the freedb site is gnudb.gnudb.org

- Add the following advanced setting:
  FreedbSiteAddress
- Use this setting in CUETools and CUERipper
- Remove the application setting MAIN_FREEDB_SITEADDRESS from CUERipper
  in favor of the new setting
- Update CUERipper Options:
  Add new setting "Freedb site address"
  Increase vertical size of Options window
  Use 32x32 cog.ico instead of 16x16 to improve quality in the task bar
- Update FreedbHelper.cs, to allow setting the site address also in
  Submit()
- Show the Freedb site address in the title of the
  "Submit to freedb" window. Resolves #157

If desired, the new setting FreedbSiteAddress can be used to disable
retrieving freedb metadata entries. It can be set to 0.0.0.0 for this
purpose.