Make sure output template contains extension .cue
The output template needs to contain an extension, which is
typically ".cue". This is relevant in cases, where e.g. the album or
artist contains a dot. If the default extension ".cue" has been
removed, then the text after the dot will be treated as the extension,
which can lead to undesired behavior.

- Check, if the output template contains ".cue" at the end and add it
  if missing.
- The check for ".cue" is done after editing the output template in
  CUERipper or CUETools and also when loading the settings file for the
  case that it has been edited.
- Resolves:
  Weird results when there are dots in filenames
    https://hydrogenaud.io/index.php?topic=122028.0
  Fallback when user doesn't add extension to template
    https://hydrogenaud.io/index.php?topic=118915.msg980706#msg980706