Move MAC_SDK here from git submodule
So far, a git submodule has been used for MAC_SDK:
https://github.com/gchudov/MAC_SDK
The MAC_SDK version used in CUETools is 4.33.
Until now, Monkey’s Audio has not been using a distributed version
control system and provides the SDK for developers as zip files found
here: https://monkeysaudio.com/files/
However, the released files are removed from time to time. The Internet
Archive remains as a possibility to continue accessing these files.
This is a preparation for future updates of MAC_SDK to newer versions.

- Use the released MAC_SDK zip file and store it here
- Origin of MAC_SDK_433.zip:
  http://monkeysaudio.com/files/MAC_SDK_433.zip
  SHA256:
  987e9a344d94019d842f685cf5a61f191fa40f000be285055c87ca671263a344
- Add the MACLibDll files, which are tailor-made for CUETools:
    MACLibDll.cpp
    MACLibDll.def
    MACLibDll.h
    MACLibDll.vcxproj
  Remark: These files have been converted to Windows (CR LF) line
  endings for conformity.
- This move reflects the contents of the submodule at commit
  gchudov/MAC_SDK@cc49a96
- Only modification in "MAC_SDK\Source\MACLibDll\MACLibDll.h:94":
  Use the intended "proc_APEDecompress_CreateEx" instead of
  "proc_APEDecompress_CreateW"
- Add info to README.md with example, how to unpack MAC_SDK_433.zip
- Update GitHub Actions accordingly